requirements.txt
black==21.5b2
flake8==3.9.2
isort==5.9.1
google-cloud-automl==2.4.2
PyGithub==1.55